3. Neither irony nor sarcasm is argument

+++ argc_argv/3-mul.c
@@ -7,7 +7,7 @@ int main(int argc, char *argv[])
 		printf("Error");
 	else 
 	{
-		printf("%d", argv[1] * argv[2]);
+		printf("%d", (int)(argv[1] * argv[2]));
 	}
 
 	return (0);
